+++ src/Math/Constraint/Constraint.hpp
@@ -9,7 +9,7 @@ class Constraint {
 	public:
 		Constraint(ConstraintId id);
 		ConstraintId getId() const;
-		virtual float getValue() const = 0;
+		virtual double getValue() const = 0;
 		virtual std::vector<std::weak_ptr<Parameter>> getDependentParameters() const = 0;// Returns list of parameters which are guaranteed to not return a null derivative
-		virtual float getValueDerivative(const Parameter& p) const = 0;
+		virtual double getValueDerivative(const Parameter& p) const = 0;
 };
